requirements.txt
Pillow
logging
tornado
requests
